+++ src/datasources/alarms/query-type-handlers/list-alarms/ListAlarmsQueryHandler.ts
@@ -1,13 +1,34 @@
-import { DataFrameDTO, DataQueryRequest, LegacyMetricFindQueryOptions, MetricFindValue } from '@grafana/data';
+import { DataFrameDTO, DataQueryRequest, DataSourceInstanceSettings, LegacyMetricFindQueryOptions, MetricFindValue } from '@grafana/data';
 import { ListAlarmsQuery } from '../../types/ListAlarms.types';
-import { AlarmsVariableQuery } from '../../types/types';
+import { AlarmsVariableQuery, QueryAlarmsRequest } from '../../types/types';
 import { AlarmsQueryHandlerCore } from '../AlarmsQueryHandlerCore';
+import { DEFAULT_QUERY_EDITOR_DESCENDING, QUERY_EDITOR_MAX_TAKE, QUERY_EDITOR_MIN_TAKE } from 'datasources/alarms/constants/AlarmsQueryEditor.constants';
 import { defaultListAlarmsQuery } from 'datasources/alarms/constants/DefaultQueries.constants';
+import { Alarm } from 'datasources/alarms/types/types';
+import { BackendSrv, getBackendSrv, getTemplateSrv, TemplateSrv } from '@grafana/runtime';
+import { User } from 'shared/types/QueryUsers.types';
+import { UsersUtils } from 'shared/users.utils';
 
 export class ListAlarmsQueryHandler extends AlarmsQueryHandlerCore {
   public readonly defaultQuery = defaultListAlarmsQuery;
 
-  public async runQuery(query: ListAlarmsQuery, _options: DataQueryRequest): Promise<DataFrameDTO> {
+  private readonly usersUtils: UsersUtils;
+
+  public constructor(
+    instanceSettings: DataSourceInstanceSettings,
+    backendSrv: BackendSrv = getBackendSrv(),
+    templateSrv: TemplateSrv = getTemplateSrv()
+  ) {
+    super(instanceSettings, backendSrv, templateSrv);
+    this.usersUtils = new UsersUtils(this.instanceSettings, this.backendSrv);
+  }
+
+  public async runQuery(query: ListAlarmsQuery, options: DataQueryRequest): Promise<DataFrameDTO> {
+    query.filter = this.transformAlarmsQuery(options.scopedVars, query.filter);
+
+    // #AB:3449773 Map queryAlarmsData response to user-selected properties
+    await this.queryAlarmsData(query);
+
     return {
       refId: query.refId,
       name: query.refId,
@@ -17,25 +38,54 @@ export class ListAlarmsQueryHandler extends AlarmsQueryHandlerCore {
 
   public async metricFindQuery(query: AlarmsVariableQuery, options?: LegacyMetricFindQueryOptions): Promise<MetricFindValue[]> {
     try {
-      const filter = this.transformAlarmsQuery(options?.scopedVars || {}, query.filter);
-      const response = await this.queryAlarms({filter});
+      const take = query.take;
+      if (!this.isTakeValid(take)) {
+        return [];
+      } 
+      
+      const filter = this.transformAlarmsQuery(options?.scopedVars || {}, query.filter) ?? '';
+      const orderByDescending = query.descending ?? DEFAULT_QUERY_EDITOR_DESCENDING;
+      const returnMostRecentlyOccurredOnly = true;
+      const requestBody = { filter, take, orderByDescending, returnMostRecentlyOccurredOnly };
+      const alarms = await this.queryAlarmsInBatches(requestBody);
 
-      const alarmsOptions = response.alarms
-        ? response.alarms.map(alarm => ({
+      const alarmsOptions = alarms.map(alarm => ({
           text: `${alarm.displayName} (${alarm.alarmId})`,
           value: alarm.alarmId
-        }))
-        : [];
-      
-      const uniqueOptions = Array.from(
-        new Map(alarmsOptions.map(option => [option.value, option])).values()
-      );
-      
-      const sortedOptions = uniqueOptions.sort((a, b) => a.text.localeCompare(b.text));
+        }));
+
+      const sortedOptions = alarmsOptions.sort((a, b) => a.text.localeCompare(b.text));
 
       return sortedOptions;
     } catch (error) {
       return [];
+    }
+  }
+
+  private isTakeValid(take?: number): boolean {
+    return take !== undefined
+      && take >= QUERY_EDITOR_MIN_TAKE
+      && take <= QUERY_EDITOR_MAX_TAKE;
+  }
+
+  private async queryAlarmsData(alarmsQuery: ListAlarmsQuery): Promise<Alarm[]> {
+    const alarmsRequestBody: QueryAlarmsRequest = {
+      filter: alarmsQuery.filter ?? '',
+    }
+
+    return this.queryAlarmsInBatches(alarmsRequestBody);
+  }
+
+  // @ts-ignore
+  // #AB:3249477 Suppress unused method warning until properties control support is implemented
+  private async loadUsers(): Promise<Map<string, User>> {
+    try {
+      return await this.usersUtils.getUsers();
+    } catch (error) {
+      if (!this.errorTitle) {
+        this.handleDependenciesError(error);
+      }
+      return new Map<string, User>();
     }
   }
 }